requirements.txt
snowflake-connector-python==2.3.4